Start splitting up mod identity_assertion
